requirements.txt
discord.py
numpy
pytz
matplotlib
python-dotenv
pillow
virtualenv
requests
git+https://github.com/ozgur/python-firebase
opencv-python